% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dream.R
\name{dream}
\alias{dream}
\title{Launch a Macaulay2 dream session}
\usage{
dream(port = 27436L, timeout = 10)
}
\value{
\code{TRUE} invisibly
}
\description{
Launch a Macaulay2 dream session
}
\examples{

\dontrun{ requires Macaulay2 be installed

dream()

# m2 code below
if (FALSE) {

  1 + 1
  R := QQ[t,x,y,z]
  I := ideal(t^4  -  x, t^3  -  y, t^2  -  z)
  gens gb I # error?

}

}

}

