% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_n.R
\name{factor_n}
\alias{factor_n}
\title{Factor an integer into primes}
\usage{
factor_n(n, code = FALSE)
}
\arguments{
\item{n}{an integer or a polynomial}

\item{code}{logical; message code to user? (default = FALSE)}
}
\value{
a character or integer matrix with columns named
  \code{prime} and \code{power}. if all the numbers have fewer
  than 10 digits, the matrix returned contains integers;
  otherwise it contains characters.
}
\description{
Factor an integer into primes
}
\examples{

\dontrun{ requires Macaulay2 be installed and an interactive session

factor_n(2^2*3^7*5^2)
factor_n(3234432540)
factor_n(323443254223453)
factor_n(rpois(1, 1e4))
factor_n("32344325422364353453")
m2("11 * 479 * 6138607975396537")
11 * 479 * 6138607975396537

}

}

