% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal.R
\name{ideal}
\alias{ideal}
\title{Create a new ideal in Macaulay2}
\usage{
ideal(mpolyList, ring, code = FALSE)
}
\arguments{
\item{mpolyList}{list of mpolys}

\item{ring}{the referent ring in Macaulay2}

\item{code}{logical; message code to user? (default = FALSE)}
}
\value{
a reference to a Macaulay2 ideal
}
\description{
Create a new ideal in Macaulay2
}
\examples{

\dontrun{ requires Macaulay2 be installed

(myring <- ring(c("x","y"), coefring = "QQ", order = "lex"))
(myideal <- ideal(mp(c("x+y", "x^2+y^2")), myring,code=TRUE))

myideal$m2_name
myideal$ring
myideal$gens
myideal$gbgens

m2(paste0("class(", myideal$m2_name, ")"))

}
}

