% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is}
\alias{is}
\alias{is.m2}
\alias{is.ring}
\title{Macaulay2 object tests}
\usage{
is.m2(x)

is.ring(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical(1)
}
\description{
Predicate functions for Macaulay2 objects.
}
\examples{

\dontrun{ requires Macaulay2 be installed

R <- ring(c("x1", "x2", "x3"))
is.m2(R)
is.ring(R)
is.ring(10)
is.ring(mp("x+1"))

}
}

