% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\name{m2_init}
\alias{m2}
\alias{m2_init}
\alias{m2_listen_code}
\alias{start_m2}
\alias{stop_m2}
\title{Initialize, call, and stop a Macaulay2 process}
\usage{
m2_listen_code(port)

start_m2(port = 27436L)

stop_m2()

m2(code)
}
\arguments{
\item{port}{port for Macaulay2 socket}

\item{code}{Macaulay2 code}
}
\value{
m2 return value
}
\description{
Initialize, call, and stop a Macaulay2 process
}
\examples{

\dontrun{ requires Macaulay2 be installed

start_m2()

m2("1 + 1")

m2("factor 32004")

# compute a grobner basis of a polynomial ideal
(gb <- m2("
  R = QQ[a..d]
  I = ideal(a^3-b^2*c, b*c^2-c*d^2, c^3)
  G = gens gb I
"))

# parse it into an mpoly object
library(stringr)
library(mpoly)
gb <- str_sub(gb, 10, -3)
gb <- str_replace_all(gb, "\\\\*", " ")
mp(str_split(gb, ", ")[[1]])

stop_m2()



}
}

