% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{m2_matrix}
\alias{m2_matrix}
\alias{m2_matrix.}
\title{Create a new matrix in Macaulay2}
\usage{
m2_matrix(mat, ring, name, code = FALSE)

m2_matrix.(mat, ring, name, code = FALSE)
}
\arguments{
\item{mat}{a matrix}

\item{ring}{a ring containing the matrix entries}

\item{name}{the \code{m2_name} of the object, which is it's name
on the M2 side}

\item{code}{return only the M2 code? (default: \code{FALSE})}
}
\value{
a reference to a Macaulay2 ring
}
\description{
Create a new matrix in Macaulay2
}
\examples{

\dontrun{ requires Macaulay2 be installed

(mymatrix <- m2_matrix(matrix(c(1,2,3,4,5,6), nrow = 3, ncol = 2)))

mymatrix$m2_name
m2(mymatrix$m2_name)
m2(paste0("class(", mymatrix$m2_name, ")"))

ring(c("x","y","z"))
mat <- matrix(mp(c("x","y","x+y","y-2","x-3","y-z")), nrow = 2, ncol = 3)
m2_matrix(mat)



}
}

