% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{m2_print}
\alias{m2_print}
\alias{print.Matrix}
\alias{print.PolynomialRing}
\title{Print a Macaulay2 object}
\usage{
\method{print}{PolynomialRing}(x, ...)

\method{print}{Matrix}(x, ...)
}
\arguments{
\item{x}{A Macaulay2 object}

\item{...}{Ignored arguments}
}
\value{
Invisible character vector of the printed object.
}
\description{
Print a Macaulay2 object
}
\examples{

\dontrun{ requires Macaulay2 be installed

R <- ring(c("x1", "x2", "x3"))
s <- print(R)
s

m <- matrix_m2(matrix(c(1,2,3,4,5,6), nrow = 3, ncol = 2))
s <- print(m)
s

}

}

