% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_to_r.R
\name{m2_to_r}
\alias{m2_to_r}
\title{Convert a M2 object into an R object}
\usage{
m2_to_r(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{...}
}
\value{
an R object
}
\description{
Convert a M2 object into an R object
}
\examples{

\dontrun{ requires Macaulay2 be installed



}
}

