% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrix_m2}
\alias{matrix_m2}
\title{Create a new matrix in Macaulay2}
\usage{
matrix_m2(mat, ring, code = FALSE)
}
\arguments{
\item{mat}{a matrix}

\item{ring}{a ring containing the matrix entries}

\item{code}{message code to user? (default = FALSE)}
}
\value{
a reference to a Macaulay2 ring
}
\description{
Create a new matrix in Macaulay2
}
\examples{

\dontrun{ requires Macaulay2 be installed

(mymatrix <- matrix_m2(matrix(c(1,2,3,4,5,6), nrow = 3, ncol = 2)))

mymatrix$m2_name
m2(mymatrix$m2_name)
m2(paste0("class(", mymatrix$m2_name, ")"))

}

}

