% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHCpack.R
\name{phc}
\alias{phc}
\alias{solve_system}
\alias{solve_system.}
\alias{mixed_volume}
\title{PHCpack}
\usage{
solve_system(mpolyList)

solve_system.(mpolyList)

mixed_volume(mpolyList)
}
\arguments{
\item{mpolyList}{An mpolyList object}
}
\value{
(currently) the output of an m2() call (string?)
}
\description{
Call PHCpack to solve a zero-dimensional system
}
\details{
Note that \code{solve_system()} doesn't take in an input ring
because the solver only works over the complex numbers.
}
\examples{

\dontrun{ requires Macaulay2

# for this to work, you need to have modified your
# init-PHCpack.m2 file instead of changing your .bashrc
# file to establish the path of phc
# (**clarify**, maybe checkout algstat::polySolve)

(mpolyList <- mp(c("t^4 - x", "t^3 - y", "t^2 - z", "x+y+z")))
solve_system(mpolyList)
mixed_volume(mpolyList)

}
}
