% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.PolynomialRing}
\alias{print.PolynomialRing}
\title{Print a Macaulay2 ring}
\usage{
\method{print}{PolynomialRing}(x, ...)
}
\arguments{
\item{x}{A \code{PolynomialRing} object}

\item{...}{Ignored arguments}
}
\value{
Invisible character vector of the printed object.
}
\description{
Print a Macaulay2 ring
}
\examples{

\dontrun{ requires Macaulay2 be installed

ring(c("x1", "x2", "x3"))
s <- print(ring(c("x1", "x2", "x3")))
s

}

}

