% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.R
\name{rm_m2}
\alias{rm_m2}
\title{Clears a Macaulay2 variable}
\usage{
rm_m2(x)
}
\arguments{
\item{x}{the M2 object to clear}
}
\value{
Invisible the object passed in.
}
\description{
Clears a Macaulay2 variable
}
\examples{

\dontrun{ requires Macaulay2 be installed

R <- ring(c("x1", "x2", "x3"))
m2(paste("class", R$m2_name))

rm_m2(R)

m2(paste("class", R$m2_name))

}

}

