% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str.PolynomialRing}
\alias{str.PolynomialRing}
\title{Give the structure of a Macaulay2 ring}
\usage{
\method{str}{PolynomialRing}(object, ...)
}
\arguments{
\item{object}{A \code{PolynomialRing} object}

\item{...}{Ignored arguments}
}
\value{
Invisible the object passed in.
}
\description{
Give the structure of a Macaulay2 ring
}
\examples{

\dontrun{ requires Macaulay2 be installed

R <- ring(c("x1", "x2", "x3"))
str(R)

}

}

