% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_m2.R
\name{str_m2}
\alias{str_m2}
\title{Give the structure of a Macaulay2 ring}
\usage{
str_m2(object, ...)
}
\arguments{
\item{object}{An \code{m2} object}

\item{...}{...}
}
\value{
Invisible the object passed in.
}
\description{
Give the structure of a Macaulay2 ring
}
\examples{

\dontrun{ requires Macaulay2

a <- m2("1")

R <- ring(c("x1", "x2", "x3"))
str_m2(R)
str_m2.default(R)

}

}
