% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_n.R
\name{factor_n}
\alias{factor_n}
\title{Factor an integer into primes}
\usage{
factor_n(n, code = FALSE, gmp = FALSE)
}
\arguments{
\item{n}{an integer or a polynomial}

\item{code}{logical; message code to user? (default: \code{FALSE})}

\item{gmp}{logical; use multiple precision arithmetic? see
\code{\link{as.bigz}}}
}
\value{
if \code{gmp = FALSE}, a character or integer matrix with
  columns named \code{prime} and \code{power}. if all the numbers
  have fewer than 10 digits, the matrix returned contains
  integers; otherwise it contains characters. if \code{gmp =
  TRUE}, a big integer matrix.
}
\description{
Factor an integer into primes
}
\examples{

\dontrun{ requires Macaulay2 be installed

##### basic usage
########################################

2^2 * 3^7 * 5^2 # = 218700
factor_n(218700)
factor_n.(218700)


##### other input types
########################################
(m2_pointer <- m2.("218700"))
factor_n(m2_pointer)
factor_n.(m2_pointer)




factor_n(3234432540)
factor_n(323443254223453)
factor_n(rpois(1, 1e4))

# notes on big integers
factor_n("32344325422364353453") # parses improperly; int too big
m2_parse(factor_n.("32344325422364353453")$ext_str)
# factor_n("32344325422364353453", gmp = TRUE)
m2("11 * 479 * 6138607975396537")
11 * 479 * 6138607975396537

}

}

