% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_poly.R
\name{factor_poly}
\alias{factor_poly}
\title{Factor a polynomial}
\usage{
factor_poly(mpoly, ring, code = FALSE)
}
\arguments{
\item{mpoly}{a character parseable by \link{\code{mp}}, an
\code{mpoly} object, or a pointer to a polynomial in M2}

\item{ring}{a \link{\code{ring}} object or a pointer to such an
object}

\item{code}{return only the M2 code? (default: \code{FALSE})}
}
\value{
a string representation of the factored polynomial.
}
\description{
Factor a polynomial
}
\examples{

\dontrun{ requires Macaulay2 be installed and an interactive session

(QQxy <- ring(c("x","y"), "QQ"))
factor_poly("x^4 - y^4", QQxy)

p <- mp("x^4 - y^4")
factor_poly.(p, QQxy)
factor_poly(p, QQxy)
factor_poly(p, QQxy, code = TRUE)
mp("(x-y) (x+y) (x^2+y^2)")


(QQxyz <- ring(c("x","y", "z"), "QQ", "lex"))
(p <- mp("(x^2 - y) (x^2 + y) (x + y)^2 (x - z)^2"))
factor_poly.(p, QQxyz)
factor_poly(p, QQxyz)

(p <- mp("(x-1)^3 (y-1)^3"))
factor_poly.(p, QQxyz)
factor_poly(p, xyz)

}

}

