% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_poly.R
\name{factor_poly}
\alias{factor_poly}
\title{Factor a polynomial}
\usage{
factor_poly(mpoly, ring, code = FALSE)
}
\arguments{
\item{mpoly}{of class mpoly}

\item{ring}{of class PolynomialRing}

\item{code}{logical; message code to user? (default = FALSE)}
}
\value{
a string representation of the factored polynomial.
}
\description{
Factor a polynomial
}
\examples{

\dontrun{ requires Macaulay2 be installed and an interactive session

(myring <- ring(c("x","y"), coefring = "QQ", order = "lex"))
p <- mp("x^4 - y^4")
factor_poly(p, myring)


}

}

