% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb.R
\name{gb}
\alias{gb}
\title{Compute a Grobner basis with Macaulay2}
\usage{
gb(mpolyList, ring = NA)
}
\arguments{
\item{mpolyList}{an mpolyList object}

\item{ring}{ring}
}
\value{
an mpolyList object
}
\description{
Compute a Grobner basis with Macaulay2
}
\examples{

\dontrun{ requires Macaulay2 be installed

( p1 <- mp("t^4 - x") )
( p2 <- mp("t^3 - y") )
( p3 <- mp("t^2 - z") )
( ms <- mpolyList(p1, p2, p3) )
gb(ms)

# easier
gb(mp(c("t^4 - x", "t^3 - y", "t^2 - z")))

}
}

