% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb.R
\name{gb}
\alias{gb}
\title{Compute a Grobner basis with gb}
\usage{
gb(mpolyList, ring = NA, vars = vars(mpolyList))
}
\arguments{
\item{mpolyList}{an mpolyList object}

\item{ring}{ring}

\item{vars}{vars}
}
\value{
an mpolyList object
}
\description{
Compute a Grobner basis with gb
}
\examples{

\dontrun{ requires Macaulay2 be installed

start_m2()

# compute a grobner basis of a polynomial ideal
(gb <- m2("
  R = QQ[a..d]
  I = ideal(a^3-b^2*c, b*c^2-c*d^2, c^3)
  G = gens gb I
"))

# parse it into an mpoly object
library(stringr)
gb <- str_sub(gb, 10, -3)
gb <- str_replace_all(gb, "\\\\*", " ")
mpoly::mp(str_split(gb, ", ")[[1]])

stop_m2()

}
}

