% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal.R
\name{ideal}
\alias{ideal}
\alias{ideal.}
\alias{print.m2_ideal}
\title{Create a new ideal in Macaulay2}
\usage{
ideal(mpolyList, ring, code = FALSE, ...)

ideal.(mpolyList, ring, code = FALSE, ...)

\method{print}{m2_ideal}(x, ...)
}
\arguments{
\item{mpolyList}{list of mpolys}

\item{ring}{the referent ring in Macaulay2}

\item{code}{return only the M2 code? (default: \code{FALSE})}

\item{...}{...}

\item{x}{(for \code{print}) x}
}
\value{
a reference to a Macaulay2 ideal
}
\description{
Create a new ideal in Macaulay2
}
\examples{

\dontrun{ requires Macaulay2 be installed

(QQxy <- ring(c("x","y"), coefring = "QQ"))
ideal(mp(c("x+y", "x^2+y^2")), QQxy)
ideal(mp(c("x+y", "x^2+y^2")), QQxy, code = TRUE)

}
}

