% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal.R
\name{ideal}
\alias{ideal}
\alias{ideal.}
\alias{print.m2_ideal}
\alias{radical}
\alias{radical.}
\title{Create a new ideal in Macaulay2}
\usage{
ideal(x, ring, raw_chars = FALSE, code = FALSE, ...)

ideal.(x, ring, raw_chars = FALSE, code = FALSE, ...)

\method{print}{m2_ideal}(x, ...)

radical(ideal, code = FALSE, ...)

radical.(ideal, code = FALSE, ...)
}
\arguments{
\item{x}{a listing of polynomials. several formats are accepted,
see examples.}

\item{ring}{the referent ring in Macaulay2}

\item{raw_chars}{if \code{TRUE}, the character vector will not be
parsed by \code{\link{mp}}, saving time (default:
\code{FALSE}). the down-side is that the strings must be
formated for M2 use directly, as opposed to for
\code{\link{mp}}. (e.g. \code{"x*y+3"} instead of \code{"x y +
3"})}

\item{code}{return only the M2 code? (default: \code{FALSE})}

\item{...}{...}

\item{ideal}{an ideal object of class \code{m2_ideal} or
\code{m2_ideal_pointer}}
}
\value{
a reference to a Macaulay2 ideal
}
\description{
Create a new ideal in Macaulay2
}
\examples{

\dontrun{ requires Macaulay2

(QQxy <- ring(c("x","y"), coefring = "QQ"))
ideal(   c("x+y", "x^2+y^2") , QQxy)
ideal(mp(c("x+y", "x^2+y^2")), QQxy)

ideal(mp(c("x+y", "x^2+y^2")), QQxy, code = TRUE)

(QQxy. <- ring.(c("x","y"), coefring = "QQ"))
ideal(   c("x+y", "x^2+y^2") , QQxy.)
ideal(mp(c("x+y", "x^2+y^2")), QQxy.)

I  <- ideal (c("x+y", "x^2+y^2"), QQxy.)
I. <- ideal.(c("x+y", "x^2+y^2"), QQxy.)
is.m2_ideal(I)
is.m2_ideal(I.)
is.m2_ideal_pointer(I)
is.m2_ideal_pointer(I.)

# radical
QQxy <- ring(c("x","y"), "QQ")
I <- ideal(c("(x^2 + 1)^2 y", "y + 1"))
radical(I)
radical.(I)

}
}
