% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is}
\alias{is}
\alias{is.m2}
\alias{is.m2_pointer}
\alias{is.ring}
\alias{is.m2_polynomialring}
\alias{is.m2_polynomialring_pointer}
\alias{is.m2_grobner_basis}
\alias{is.m2_ideal}
\alias{is.m2_ideal_pointer}
\alias{is.m2_ideal_list}
\alias{is.m2_ideal_list_pointer}
\alias{is.m2_module}
\alias{is.m2_option}
\alias{is.m2_matrix}
\alias{is.m2_matrix_pointer}
\alias{is.m2_list}
\alias{is.m2_array}
\alias{is.m2_sequence}
\title{Macaulay2 object tests}
\usage{
is.m2(x)

is.m2_pointer(x)

is.ring(x)

is.m2_polynomialring(x)

is.m2_polynomialring_pointer(x)

is.m2_grobner_basis(x)

is.m2_ideal(x)

is.m2_ideal_pointer(x)

is.m2_ideal_list(x)

is.m2_ideal_list_pointer(x)

is.m2_module(x)

is.m2_option(x)

is.m2_matrix(x)

is.m2_matrix_pointer(x)

is.m2_list(x)

is.m2_array(x)

is.m2_sequence(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical(1)
}
\description{
Predicate functions for Macaulay2 objects.
}
\examples{

\dontrun{ requires Macaulay2

R <- ring(c("x1", "x2", "x3"))
is.m2(R)
is.ring(R)
is.ring(10)
is.ring(mp("x+1"))

}
}
