% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\name{m2_call}
\alias{m2}
\alias{m2.}
\alias{m2_call}
\alias{m2_listen_code}
\alias{reset_m2}
\title{Call and reset a Macaulay2 process}
\usage{
m2_listen_code(port, timeout)

reset_m2(port = 27436L, timeout = 10)

m2(code, timeout = -1)

m2.(code, timeout = -1)
}
\arguments{
\item{port}{port for Macaulay2 socket}

\item{timeout}{number of seconds before aborting}

\item{code}{Macaulay2 code}
}
\value{
m2 return value
}
\description{
Call and reset a Macaulay2 process
}
\examples{

\dontrun{ requires Macaulay2 be installed

m2("1 + 1")

m2("factor 32004")

# compute a grobner basis of a polynomial ideal
m2("
  R = QQ[a..d]
  I = ideal(a^3-b^2*c, b*c^2-c*d^2, c^3)
  G = gens gb I
")

m2("a = 1 + 1")
m2("a")
reset_m2()
m2("a")

}
}

