% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_utility.R
\name{m2_utility}
\alias{m2_exists}
\alias{m2_getwd}
\alias{m2_ls}
\alias{m2_meta}
\alias{m2_name}
\alias{m2_rm}
\alias{m2_utility}
\title{Utility tools for M2}
\usage{
m2_name(x)

m2_meta(x, attr)

m2_exists(name)

m2_ls()

m2_rm()

m2_getwd()
}
\arguments{
\item{name}{name of object}
}
\description{
Utility tools for M2
}
\examples{

\dontrun{ requires Macaulay2 be installed

m2("a = 5")
m2_ls()
m2_exists("a")
m2("b = 1")
m2_exists(c("a","b","c"))

m2_getwd()


}
}

