% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2r-package.R
\docType{package}
\name{m2r}
\alias{m2r}
\alias{package-m2r}
\title{Macaulay2 in R}
\description{
m2r provides a persistent interface to Macaulay2
(\url{http://www.math.uiuc.edu/Macaulay2/}) and front-end tools facilitating
its use in the R ecosystem. For details, see vignette("m2r").
}
