% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring.R
\name{ring}
\alias{ring}
\alias{ring.}
\title{Create a new ring in Macaulay2}
\usage{
ring(vars, coefring = m2_coefrings(), order = m2_termorders(),
  code = FALSE)

ring.(vars, coefring = m2_coefrings(), order = m2_termorders(),
  code = FALSE)
}
\arguments{
\item{vars}{vector of variable names}

\item{coefring}{coefficient ring (default: \code{"CC"})}

\item{order}{a term order (default: \code{"grevlex"})}

\item{code}{message code to user? (default: \code{FALSE})}
}
\value{
a reference to a Macaulay2 ring
}
\description{
Create a new ring in Macaulay2
}
\examples{

\dontrun{ requires Macaulay2 be installed

(myring <- ring(c("x1","x2","x3","y"), coefring = "QQ", order = "lex"))

myring$m2_name
myring$vars
myring$coefring
myring$order

m2_code <- sprintf("class(\%s)", myring$m2_name)
m2(m2_code)



r. <- ring.(c("x1","x2","x3","y"), coefring = "QQ", order = "lex")
m2(r.$m2_name) # = r.$ext_str

}

}

