% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring.R
\name{ring}
\alias{ring}
\title{Create a new ring instance in Macaulay2}
\usage{
ring(vars, coefring = "CC", order = "lex")
}
\arguments{
\item{vars}{list of variables}

\item{coefring}{coefficient ring}
}
\value{
a reference to a Macaulay2 ring
}
\description{
Create a new ring instance in Macaulay2
}
\examples{

\dontrun{ requires Macaulay2 be installed

( p1 <- mp("t^4 - x") )
( p2 <- mp("t^3 - y") )
( p3 <- mp("t^2 - z") )
( ms <- mpolyList(p1, p2, p3) )
gb(ms)

}
}

