% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring.R
\name{ring}
\alias{ring}
\title{Create a new ring in Macaulay2}
\usage{
ring(vars, coefring = c("CC", "RR", "QQ", "ZZ"), order = c("grevlex", "lex",
  "glex"), code = FALSE)
}
\arguments{
\item{vars}{vector of variable names}

\item{coefring}{coefficient ring}

\item{order}{a term order}

\item{code}{message code to user? (default = FALSE)}
}
\value{
a reference to a Macaulay2 ring
}
\description{
Create a new ring in Macaulay2
}
\examples{

\dontrun{ requires Macaulay2 be installed

(myring <- ring(c("x1","x2","x3","y"), coefring = "QQ", order = "lex"))

myring$m2name
myring$vars
myring$coefring
myring$order

m2(paste0("class(", myring$m2name, ")"))

}

}

