% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf.R
\name{snf}
\alias{snf}
\title{Smith normal form}
\usage{
snf(mat, code = FALSE)
}
\arguments{
\item{mat}{a matrix (integer entries)}

\item{code}{logical; message code to user? (default = FALSE)}
}
\value{
a list of integer matrices D, P, and Q
}
\description{
For an integer matrix M, this computes the matrices D, P, and Q
such that \emph{D = PMQ}, which can be seen as an analogue of the
singular value decomposition.
}
\examples{

\dontrun{ requires Macaulay2 be installed and an interactive session

M <- matrix(c(
   2,  4,   4,
  -6,  6,  12,
  10, -4, -16
), nrow = 3, byrow = TRUE)
(mats <- snf(M))
P <- mats$P; D <- mats$D; Q <- mats$Q
P \%*\% M \%*\% Q # = D

}

}

